OSX related naming changes

+++ include/azul/ipc/SharedMemory.hpp
@@ -11,12 +11,12 @@ namespace azul
         class SharedMemory final
         {
         public:
-            explicit SharedMemory(std::string const& name, std::size_t const size, bool const is_owner);
+            explicit SharedMemory(std::string const& name, std::uint64_t const size, bool const is_owner);
             SharedMemory();
             ~SharedMemory();
 
             void* Address() const;
-            std::size_t Size() const;
+            std::uint64_t Size() const;
 
             SharedMemory& operator=(SharedMemory const&) = delete;
             SharedMemory(SharedMemory const&) = delete;
